requirements.txt
python-dotenv
requests
websocket-client
git+https://github.com/zigpy/zha-device-handlers.git
watchdog
